Refactoring
The abstract classes approach was clumsy to implement.  Thus by using
generic types and one interface I managed to reduce the classes that
needs to be coded from three classes to two classes.

These two classes do not implement/extend any complicated classes so the
implementation is a bit cleaner. The enumeration class is ready for
refactoring and it is NOT tested. In the next commit I will remove the
classes I can get rid of off once I have tested these new
implementations a bit more.